Rename utils file
